Models reading from config.yml

+++ packages/back-end/src/models/DataSourceModel.ts
@@ -12,6 +12,7 @@ import {
   testDataSourceConnection,
 } from "../services/datasource";
 import uniqid from "uniqid";
+import { usingFileConfig, getConfigDatasources } from "../init/config";
 
 const dataSourceSchema = new mongoose.Schema({
   id: {
@@ -76,28 +77,60 @@ const dataSourceSchema = new mongoose.Schema({
     },
   },
 });
-export type DataSourceDocument = mongoose.Document & DataSourceInterface;
+type DataSourceDocument = mongoose.Document & DataSourceInterface;
 
 const DataSourceModel = mongoose.model<DataSourceDocument>(
   "DataSource",
   dataSourceSchema
 );
 
-export function getOrganizationsWithDatasources() {
-  return DataSourceModel.distinct("organization");
+function toInterface(doc: DataSourceDocument): DataSourceInterface {
+  if (!doc) return null;
+  return doc.toJSON();
 }
-export function deleteDatasourceById(id: string) {
-  DataSourceModel.deleteOne({
+
+export async function getDataSourcesByOrganization(organization: string) {
+  // If using config.yml, immediately return the list from there
+  if (usingFileConfig()) {
+    return getConfigDatasources(organization);
+  }
+
+  return (
+    await DataSourceModel.find({
+      organization,
+    })
+  ).map(toInterface);
+}
+export async function getDataSourceById(id: string, organization: string) {
+  // If using config.yml, immediately return the from there
+  if (usingFileConfig()) {
+    return (
+      getConfigDatasources(organization).filter((d) => d.id === id)[0] || null
+    );
+  }
+
+  const doc = await DataSourceModel.findOne({
     id,
   });
+
+  if (doc && doc.organization !== organization) {
+    throw new Error("You do not have access to that datasource");
+  }
+
+  return toInterface(doc);
 }
-export async function getDataSourcesByOrganization(organization: string) {
-  return await DataSourceModel.find({
-    organization,
-  });
+
+export async function getOrganizationsWithDatasources(): Promise<string[]> {
+  if (usingFileConfig()) {
+    return [];
+  }
+  return await DataSourceModel.distinct("organization");
 }
-export async function getDataSourceById(id: string) {
-  return await DataSourceModel.findOne({
+export async function deleteDatasourceById(id: string) {
+  if (usingFileConfig()) {
+    throw new Error("Cannot delete. Data sources managed by config.yml");
+  }
+  await DataSourceModel.deleteOne({
     id,
   });
 }
@@ -109,6 +142,10 @@ export async function createDataSource(
   params: DataSourceParams,
   settings?: DataSourceSettings
 ) {
+  if (usingFileConfig()) {
+    throw new Error("Cannot add. Data sources managed by config.yml");
+  }
+
   const id = uniqid("ds_");
 
   if (type === "google_analytics") {
@@ -134,5 +171,23 @@ export async function createDataSource(
   await testDataSourceConnection(datasource);
   const model = await DataSourceModel.create(datasource);
 
-  return model;
+  return toInterface(model);
+}
+
+export async function updateDataSource(
+  id: string,
+  updates: Partial<DataSourceInterface>
+) {
+  if (usingFileConfig()) {
+    throw new Error("Cannot update. Data sources managed by config.yml");
+  }
+
+  await DataSourceModel.updateOne(
+    {
+      id,
+    },
+    {
+      $set: updates,
+    }
+  );
 }
